fix: correct component name, rename file to match import/name
refactor: remove unnec ref()
